feat: Create content section for dashboard
This commit adds a new content section for the dashboard. The section includes a variety of widgets, such as cards and tables, to display important metrics and data to users. The table are customizable and can be rearranged to suit the user's needs.

Files added:
- src/assets/users/usser-details.svg
- src/components/loading/index.tsx
- src/components/loading/index.scss
- src/components/userDetails/index.tsx
- src/components/loading/index.scss
- src/components/userCard/index.tsx
- src/components/userCard/index.scss
- src/context/appProvider.tsx
- src/context/index.tsx
src/hooks/useFetchId.ts

Files modified:
- src/hooks/useFetch.ts
- src/style/index.ts
- src/App.tsx
- src/ap/dashboard/index.scss
- src/app/dashboard.tsx